requirements.txt
requests
sqlalchemy
mysql-connector-python
python-dotenv
pandas
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
google-generativeai>=0.3.0
langgraph>=0.0.40
langchain-core>=0.1.0
reportlab>=4.0.0
aiohttp>=3.8.0
requests>=2.31.0
Pillow>=10.0.0